Small updates.
Update create and edit views and forms for Equipment, FieldNote (and
later, other site-related classes) so that the site field is filled in
when creation is started from a site page and left blank when done
from a list page, and add a method to those classes to enable or
disable editing of the site field on a per-class basis.
Add a site_visitors field to FieldNotes.

Rename field Equipment.location to Equipment.site to match other
classes.
Correct a typo in EquipmentViewsMixin.get_context_data().